Fixes for MinGW-w64 native Windows build
As discussed in https://sourceforge.net/p/openipmi/bugs/92/ here is my
first pull request with changes needed to fix some things to be able to
build for native Windows using MinGW-w64.

Fixes include solutions to resolve the following:

 * Issues related to the fact that sizeof(long) is not the same as
   sizeof(void*). Solution: use intptr_t and uintptr_t from #include
   <stdint.h> as integer datatypes that are the same size as a pointer.

 * Setting a socket to non-blocking mode is done differently on Windows.

 * inet_ntop and inet_pton are not found if _WIN32_WINNT is not at least 0x0600

With these fixes make -Cutils compiles without errors on MinGW-w64 GCC
10.2.0 (I use my personal build from http://winlibs.com/).

+++ lib/rakp.c
@@ -33,6 +33,7 @@
 
 #include <config.h>
 
+#include <stdint.h>
 #include <string.h>
 
 #include <OpenIPMI/ipmi_conn.h>
@@ -138,7 +139,7 @@ handle_rakp4(ipmi_con_t *ipmi, ipmi_msgi_t *rspi)
 {
     ipmi_msg_t  *msg = &rspi->msg;
     rakp_info_t *info = rspi->data1;
-    int         addr_num = (long) rspi->data4;
+    int         addr_num = (intptr_t) rspi->data4;
     int         rv;
     uint32_t    session_id;
 
@@ -220,7 +221,7 @@ handle_rakp2(ipmi_con_t *ipmi, ipmi_msgi_t *rspi)
 {
     ipmi_msg_t    *msg = &rspi->msg;
     rakp_info_t   *info = rspi->data1;
-    int           addr_num = (long) rspi->data4;
+    int           addr_num = (intptr_t) rspi->data4;
     int           rv;
     uint32_t      session_id;
     int           err = 0;
